move AES-NI code to internal package
This change moves the platform specific code for AES hardware
support detection to a separate internal package.
