+++ internal/cpu/aesni_generic.go
@@ -12,9 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// +build amd64,!gccgo,!appengine
+// +build !amd64 gccgo appengine
 
-package sio
+package cpu
 
-//go:noescape
-func hasAESNISupport() bool
+func hasAESNISupport() bool { return false }
